use regex to capture cabal key instead of URL
apparently the new URL() behaviour is unpredictable. chrome, node, and
firefox all return different values when a cabal key is passed in.

for node the cabal key can be found in `uri.host`, in chromium/electron
it's found at `uri.hostname`, for firefox both of those are empty, but
`uri.pathname` gives the cabal key prepended with two slashes (but that
doesn't work in node...)

until the IANA change lands / has a standardizing effect, let's just use
a regex to capture it.

fixes #88