add preferredPort (default 13331) to core opts
add a preferredPort to listen for hyperswarm connections on. the
preferredPort is smart enough that if it is blocked, hyperswarm will try
preferredPort+1, preferredPort+2 before giving up. see the hyperswarm code:
https://github.com/hyperswarm/network/blob/ad41b0c2ac19dbf92865e481282ba6ac6e513d57/index.js#L150-L162

this allows people with highly configured firewalls to port forward a
single port for cabal, which they can themselves configure in their
clients.

fix #98